refactor: move pages to seperate folder
